+++ cloud/readme.md
@@ -1,0 +1,30 @@
+# 使用fastapi完成的适合本模组的服务端
+
+##### 如何使用：
+
+    pip install fastapi uvicorn
+
+    python local.py
+
+    程序会默认开放于本地的8235端口，使用http协议访问。
+
+##### 配置参数:
+
+    可以在local.py文件开头更改密钥和端口。
+
+    默认使用safekey =  “dinocekey”  port = 8235
+
+    程序会在local.py所在位置创建save文件夹，存档以二进制文件保存在此。
+
+/save
+  └─ [存档名称]
+      ├─ [存档名称].save  # 实际文件
+      └─ time.txt        # 时间戳记录
+
+##### 注意事项：
+
+    如果你希望跨平台，跨设备使用云存档，设置相同的存档名称即可，但是即使使用相同的存档名称，对于从未上传过存档的设备，或是你刚刚更改完云存档设置的任何内容，必须要首先上传一次存档，从未上传过存档的游戏下载存档会失败。如果你需要和我一样在多个设备上游玩，第一次使用云存档之前，你仍然需要使用游戏自带的功能转移游戏数据，然后上传一次存档，此后就可以正常使用云存档功能了。
+
+##### Todo：
+
+    对于从未上传过存档的客户端，同步存档的时候会出现”Too much data for declared Content-Length“错误，尚不明确为什么会有这个问题，有时间再做吧。
